fix: low-severity batch across kernel, libc, crypto, media and net
Eighteen confirmed findings from the 2026-09-16 audit (docs/bugs/README.md
has the full list). Highlights: ktime CPU accounting gains LRU slot
eviction (65th pid used to fold into idle silently); modelf validates that
section/symbol string tables end in NUL and compares bounded (a truncated
.ko walked past the image); mixer/capture buffer-ms computed in 64-bit;
pipe read/write honour descriptor direction; ptrace ATTACH of a zombie
returns SRCH instead of a 2-second timeout; SYS_READ_FILE sizes its buffer
from the file before allocating; x25519 and ML-DSA wipe secret intermediates;
eval_string refuses on OOM instead of printing None; login stops refunding
tries on empty usernames; terminal paste buffer matches the clipboard cap;
vorbis/reasm/h265 entry-point wrap fixes.

+++ c/kernel/audio/capture.c
@@ -443,7 +443,7 @@ void snd_cap_report(void)
             g_capdev->period_bytes / (g_capdev->channels * 2u),
             (g_capdev->period_bytes / (g_capdev->channels * 2u)) * 1000u / g_capdev->rate,
             g_capdev->periods,
-            (g_capdev->period_bytes / (g_capdev->channels * 2u)) * 1000u * g_capdev->periods / g_capdev->rate,
+            (unsigned)((uint64_t)(g_capdev->period_bytes / (g_capdev->channels * 2u)) * 1000u * g_capdev->periods / g_capdev->rate),
             g_capdev->irq_mode == 3 ? "msix" : g_capdev->irq_mode == 2 ? "msi"
           : g_capdev->irq_mode == 1 ? "intx" : "polled");
 }
